refactor(user): 프로필 업데이트 정책 분리
