+++ src/users/application/usecases/update-me.usecase.ts
@@ -1,11 +1,12 @@
 import { Inject, Injectable } from '@nestjs/common';
-import {
-  USERS_REPOSITORY,
-  UpdateAuthAccountProfileParams,
-} from '../../domain/users.repository';
+import { USERS_REPOSITORY } from '../../domain/users.repository';
 import type { UsersRepository } from '../../domain/users.repository';
 import { mapMeResponse } from '../policies/map-me-response.policy';
 import { UsersError } from '../errors/users.error';
+import {
+  buildUpdateMeParams,
+  resolveCurrentAuthAccount,
+} from '../policies/update-me.policy';
 
 export type UpdateMeInput = {
   displayName?: string | null;
@@ -26,19 +27,8 @@ export class UpdateMeUseCase {
       throw new UsersError('NOT_FOUND', '사용자를 찾을 수 없습니다.');
     }
 
-    const currentAccount = user.authAccounts.find(
-      (item) => item.id === accountId,
-    );
-
-    if (!currentAccount) {
-      throw new UsersError('NOT_FOUND', '계정 정보를 찾을 수 없습니다.');
-    }
-
-    if (input.displayName === null) {
-      throw new UsersError('BAD_REQUEST', 'displayName은 null일 수 없습니다.');
-    }
-
-    const updateParams = this.toUpdateParams(input);
+    resolveCurrentAuthAccount(user, accountId);
+    const updateParams = buildUpdateMeParams(input);
 
     if (Object.keys(updateParams).length === 0) {
       return mapMeResponse(user, accountId);
@@ -57,19 +47,5 @@ export class UpdateMeUseCase {
     };
 
     return mapMeResponse(mergedUser, accountId);
-  }
-
-  private toUpdateParams(input: UpdateMeInput): UpdateAuthAccountProfileParams {
-    const params: UpdateAuthAccountProfileParams = {};
-
-    if (input.displayName !== undefined && input.displayName !== null) {
-      params.displayName = input.displayName;
-    }
-
-    if (input.avatarUrl !== undefined) {
-      params.profileImageUrl = input.avatarUrl;
-    }
-
-    return params;
   }
 }
